+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Embed defined namespace
+- Expand imported but unused namespaces
 
 ## [4.0.2] - 2022-01-23
 - Downgrade libraries SourceExpander.Embedder.Testing